bootloader: use arm-linux-androidkernel- for l-loader
The difference between arm-linux-androidkernel- and
arm-linux-androideabi- is 'ld', bfd vs. gold.  We have seen problem with
ld.gold building l-loder, that is the 'fip' image section will not be at
offset 0x3f000, but shifted 62 bytes higher for some reason we do not
understand.  Let's use ld.bfd which is more like the gnueabi version and
doesn't have the problem seen with ld.gold.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>